requirements.txt
Flask
requests
nltk
scikit-learn
azure-storage
pandas